edit_constants
