+++ experiments/optimizer/GPT2/constants.py
@@ -1,0 +1,4 @@
+model_name = "flax-community/gpt2-base-thai"
+bos_token = "<|startoftext|>"
+eos_token = "<|endoftext|>"
+pad_token = "<|pad|>"
